requirements.txt
Django==3.0.5
gunicorn==19.9.0
djangorestframework==3.11.0
whitenoise==5.0.1
dj-database-url==0.5.0